Added explicit cast in GemObject::moveToBoardPosition to stop the compiler from complaining about it.

+++ GemObject.cpp
@@ -38,5 +38,5 @@ void GemObject::setLerpSpeed( float speed )
 
 void GemObject::moveToBoardPosition( Coordinates position )
 {
-	transform()->position = { position.x,position.y };
+	transform()->position = { (double)position.x,(double)position.y };
 }
